cri: trim redundant/implementation-detail tests, share fake-executable scaffolding
CdpClient.connect's unavailable-port test duplicated the deterministic
connectCdp coverage while costing ~12s per run -- traced to Deno's fetch
taking ~2.3s per refused connect against a closed "localhost" port
(measured directly; not an internal deno-cri retry). Real integration
coverage of CdpClient against the vendored deno-cri library already
exists via tests/smoke/axe/axe-transport-failclosed.test.ts, which drives
a real browser through connect/close/dropped-socket.

Dropped the no-dangling-colon-on-empty-stderr case (a cosmetic message
branch already exercised structurally by the non-empty-stderr test) and
the direct probeTimeoutMs unit test, making probeTimeoutMs private again
now that nothing outside launch.ts needs to call it directly.

In chrome-launch-flags.test.ts, replaced the second URL-gated fake-browser
fixture and its dedicated test with a one-line argv assertion on the
existing launch-arguments test, since criClient always passes the same
url: "about:blank" regardless of fixture behavior.

Extracted the repeated "write a fake Chrome .cmd/.sh wrapper executable"
boilerplate (three near-identical copies across the two files) into a
shared writeFakeExecutable helper in the new chrome-launch-fixtures.ts.
