+++ tests/unit/chrome-launch.test.ts
@@ -6,15 +6,11 @@
 
 import { assert, assertEquals, assertRejects } from "testing/asserts";
 import { join } from "path";
-import {
-  connectCdp,
-  hasPageTarget,
-  launchChrome,
-  probeTimeoutMs,
-} from "../../src/core/cri/launch.ts";
-import { CdpClient } from "../../src/command/call/axe/scan.ts";
+import { connectCdp, hasPageTarget, launchChrome } from
+  "../../src/core/cri/launch.ts";
 import { findOpenPort } from "../../src/core/port.ts";
 import { unitTest } from "../test.ts";
+import { writeFakeExecutable } from "./chrome-launch-fixtures.ts";
 
 /** Short enough that the test doesn't wait around, long enough to be real. */
 const kLaunchTimeout = 300;
@@ -24,17 +20,7 @@ const kSelfDestructMs = 2000;
 // A fake Chrome that accepts the TCP connection on its CDP port but never
 // answers on it -- genuinely silent, not merely absent, so a `fetch()`
 // against it hangs on the response rather than failing fast with connection
-// refused. Same wrapper mechanism as chrome-launch-flags.test.ts's fake
-// executable (.cmd on Windows, a shebang script on Unix, re-invoking the
-// current deno binary to run an inline script -- a wrapper is needed because
-// launchChrome() always prepends real Chrome flags as argv, which the fake
-// executable must tolerate).
-//
-// It self-exits on its own short timer rather than relying on launchChrome's
-// kill() to reach it: on Windows, kill() only reaches the direct child (the
-// .cmd's cmd.exe host), and the actual deno.exe process survives as an
-// orphan holding its inherited stderr pipe open, which would otherwise hang
-// launchChrome's close() until the timer fires anyway.
+// refused.
 async function writeSilentChromeExecutable(
   dir: string,
   port: number,
@@ -42,9 +28,9 @@ async function writeSilentChromeExecutable(
   selfDestructMs: number = kSelfDestructMs,
   exitTimestampPath?: string,
 ): Promise<string> {
-  const scriptPath = join(dir, "silent-chrome.js");
-  await Deno.writeTextFile(
-    scriptPath,
+  return writeFakeExecutable(
+    dir,
+    "silent-chrome",
     [
       `const listener = Deno.listen({ port: ${port} });`,
       "(async () => {",
@@ -70,28 +56,9 @@ async function writeSilentChromeExecutable(
         : "",
       `  Deno.exit(1);`,
       `}, ${selfDestructMs});`,
-    ].join("\n"),
+    ],
+    exitTimestampPath,
   );
-
-  const deno = Deno.execPath();
-  const allowWrite = exitTimestampPath
-    ? ` --allow-write="${exitTimestampPath}"`
-    : "";
-  if (Deno.build.os === "windows") {
-    const cmdPath = join(dir, "silent-chrome.cmd");
-    await Deno.writeTextFile(
-      cmdPath,
-      `@echo off\r\n"${deno}" run --allow-net${allowWrite} "${scriptPath}" %*\r\n`,
-    );
-    return cmdPath;
-  }
-  const shPath = join(dir, "silent-chrome.sh");
-  await Deno.writeTextFile(
-    shPath,
-    `#!/bin/sh\n"${deno}" run --allow-net${allowWrite} "${scriptPath}" "$@"\n`,
-  );
-  await Deno.chmod(shPath, 0o755);
-  return shPath;
 }
 
 unitTest(
@@ -119,36 +86,6 @@ unitTest(
           "Chrome said: Chrome blew up: missing shared library",
         ),
         `expected stderr tail after "Chrome said:" in: ${err.message}`,
-      );
-    } finally {
-      await Deno.remove(dir, { recursive: true }).catch(() => {});
-    }
-  },
-);
-
-unitTest(
-  "chrome-launch - alive-but-silent Chrome rejects, no dangling colon on empty stderr",
-  async () => {
-    const port = findOpenPort();
-    const dir = await Deno.makeTempDir({ prefix: "chrome-launch-silent-" });
-    try {
-      const fakeChrome = await writeSilentChromeExecutable(dir, port, "");
-      const err = await assertRejects(
-        () =>
-          launchChrome({ appPath: fakeChrome, port, timeout: kLaunchTimeout }),
-        Error,
-      );
-      assert(
-        err.message.includes(String(port)),
-        `expected port ${port} named in: ${err.message}`,
-      );
-      assert(
-        !err.message.includes("Chrome said:"),
-        `expected no "Chrome said:" for empty stderr in: ${err.message}`,
-      );
-      assert(
-        !err.message.trimEnd().endsWith(":"),
-        `expected no dangling colon in: ${err.message}`,
       );
     } finally {
       await Deno.remove(dir, { recursive: true }).catch(() => {});
@@ -202,18 +139,6 @@ unitTest(
   },
 );
 
-unitTest(
-  "chrome-launch - CdpClient.connect exhausts its retries with a useful message",
-  async () => {
-    const port = findOpenPort();
-    const err = await assertRejects(() => CdpClient.connect(port), Error);
-    assert(
-      err.message.includes(String(port)),
-      `expected port ${port} named in: ${err.message}`,
-    );
-  },
-);
-
 // deno-lint-ignore require-await
 unitTest(
   "chrome-launch - hasPageTarget requires target.type to be page",
@@ -234,19 +159,6 @@ unitTest(
     );
     assertEquals(hasPageTarget([]), false);
     assertEquals(hasPageTarget("not-an-array"), false);
-  },
-);
-
-// deno-lint-ignore require-await
-unitTest(
-  "chrome-launch - probeTimeoutMs never exceeds what's left of the launch budget",
-  async () => {
-    // Plenty of time left: capped at the per-attempt ceiling, not the full budget.
-    assertEquals(probeTimeoutMs(5000), 1000);
-    // Little time left: capped at what's actually left, not the per-attempt ceiling.
-    assertEquals(probeTimeoutMs(50), 50);
-    // Already past the deadline: clamps to zero rather than going negative.
-    assertEquals(probeTimeoutMs(-10), 0);
   },
 );
 
@@ -287,60 +199,34 @@ unitTest(
 
 /**
  * A fake Chrome that serves `/json/list` itself, controlling exactly what the
- * launcher's readiness poll sees -- same Deno.serve wrapper mechanism as
- * chrome-launch-flags.test.ts's fake executable. `pageAfterMs` is measured
- * from the fake server's own start, not the caller's: `[]` until then, a
- * single real page target afterward. `undefined` means never.
+ * launcher's readiness poll sees. `pageAfterMs` is measured from the fake
+ * server's own start, not the caller's: `[]` until then, a single real page
+ * target afterward. `undefined` means never.
  */
 async function writeJsonListChromeExecutable(
   dir: string,
   port: number,
   pageAfterMs: number | undefined,
 ): Promise<string> {
-  const scriptPath = join(dir, "json-list-chrome.js");
-  await Deno.writeTextFile(
-    scriptPath,
-    [
-      "const start = Date.now();",
-      `const pageAfterMs = ${
-        pageAfterMs === undefined ? "undefined" : pageAfterMs
-      };`,
-      `Deno.serve({ port: ${port}, onListen: () => {} }, (req) => {`,
-      "  const url = new URL(req.url);",
-      '  if (url.pathname !== "/json/list") {',
-      '    return new Response("", { status: 404 });',
-      "  }",
-      "  const ready = pageAfterMs !== undefined &&",
-      "    Date.now() - start >= pageAfterMs;",
-      "  const body = ready",
-      `    ? JSON.stringify([{ type: "page", webSocketDebuggerUrl: "ws://localhost:${port}/devtools/page/1" }])`,
-      '    : "[]";',
-      "  return new Response(body, { status: 200 });",
-      "});",
-      // Same Windows orphan issue as writeSilentChromeExecutable above:
-      // launchChrome's close() only kills the .cmd's cmd.exe host, and the
-      // deno.exe grandchild survives holding its inherited stderr pipe open,
-      // which would otherwise hang close()'s drain loop forever.
-      `setTimeout(() => Deno.exit(0), ${kSelfDestructMs});`,
-    ].join("\n"),
-  );
-
-  const deno = Deno.execPath();
-  if (Deno.build.os === "windows") {
-    const cmdPath = join(dir, "json-list-chrome.cmd");
-    await Deno.writeTextFile(
-      cmdPath,
-      `@echo off\r\n"${deno}" run --allow-net "${scriptPath}" %*\r\n`,
-    );
-    return cmdPath;
-  }
-  const shPath = join(dir, "json-list-chrome.sh");
-  await Deno.writeTextFile(
-    shPath,
-    `#!/bin/sh\n"${deno}" run --allow-net "${scriptPath}" "$@"\n`,
-  );
-  await Deno.chmod(shPath, 0o755);
-  return shPath;
+  return writeFakeExecutable(dir, "json-list-chrome", [
+    "const start = Date.now();",
+    `const pageAfterMs = ${
+      pageAfterMs === undefined ? "undefined" : pageAfterMs
+    };`,
+    `Deno.serve({ port: ${port}, onListen: () => {} }, (req) => {`,
+    "  const url = new URL(req.url);",
+    '  if (url.pathname !== "/json/list") {',
+    '    return new Response("", { status: 404 });',
+    "  }",
+    "  const ready = pageAfterMs !== undefined &&",
+    "    Date.now() - start >= pageAfterMs;",
+    "  const body = ready",
+    `    ? JSON.stringify([{ type: "page", webSocketDebuggerUrl: "ws://localhost:${port}/devtools/page/1" }])`,
+    '    : "[]";',
+    "  return new Response(body, { status: 200 });",
+    "});",
+    `setTimeout(() => Deno.exit(0), ${kSelfDestructMs});`,
+  ]);
 }
 
 unitTest(
